+++ src/lib/stores/currentTmpPoem.ts
@@ -1,12 +1,12 @@
 import { writable } from 'svelte/store';
 
-import { Preferences } from '@capacitor/preferences';
-import FilesystemWithPermissions from '../util/FilesystemWithPermissions';
+import { Preferences } from '$lib/plugins/Preferences';
+import { Filesystem } from '$lib/plugins/Filesystem';
 
 export const currentTmpPoemStore = async () => {
 	const store = writable<string>(
 		(
-			await FilesystemWithPermissions.readFile({
+			await Filesystem.readFile({
 				path: (await Preferences.get({ key: 'current_poem_uri' })).value as string
 			})
 		).data.toString()