+++ src/routes/+layout.svelte
@@ -1,6 +1,6 @@
 <!--
 PokeBook -- Pokeghost's poetry noteBook
-Copyright (C) 2023-2024 Pokeghost.
+Copyright (C) 2023-2024, 2026 Pokeghost.
 
 PokeBook is free software: you can redistribute it and/or modify
 it under the terms of the GNU Affero General Public License as published
@@ -17,11 +17,8 @@ along with this program. If not, see <https://www.gnu.org/licenses/>.
 -->
 
 <script lang="ts">
-	import { Capacitor } from '@capacitor/core';
-	import { StatusBar } from '@capacitor/status-bar';
 	import { Toaster } from 'svelte-french-toast';
 	import { Modals, closeModal } from 'svelte-modals';
-	import { App } from '@capacitor/app';
 
 	import { darkMode } from '$lib/stores/darkMode';
 	import { dayTheme } from '$lib/stores/dayTheme';
@@ -33,18 +30,6 @@ along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 	$: $darkMode, updateTheme();
 
-	function rgbToHex(r: number, g: number, b: number) {
-		return (
-			'#' +
-			[r, g, b]
-				.map((x) => {
-					const hex = x.toString(16);
-					return hex.length === 1 ? '0' + hex : hex;
-				})
-				.join('')
-		);
-	}
-
 	function updateTheme() {
 		document.documentElement.className = '';
 		if ($darkMode !== '') {
@@ -53,20 +38,7 @@ along with this program. If not, see <https://www.gnu.org/licenses/>.
 		} else {
 			document.documentElement.classList.add($dayTheme || 'vanilla');
 		}
-		// I'm not a big fan of this idea but it's better than an ugly empty bar so it'll do for now
-		if (Capacitor.isNativePlatform()) {
-			const backgroundColorValues = getComputedStyle(document.body)
-				.getPropertyValue('background-color')
-				.match(/rgb\((\d+),\s*(\d+),\s*(\d+)\)/);
-			const [, red, green, blue] = backgroundColorValues!.map(Number);
-			const backgroundColorHex = rgbToHex(red, green, blue);
-			StatusBar.setBackgroundColor({ color: backgroundColorHex });
-		}
 	}
-
-	App.addListener('backButton', (_) => {
-		window.history.back();
-	});
 </script>
 
 <Modals>