+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [v3.1.1] - 2026-02-01
+
+### Changed
+
+- Remove Capacitor dependencies
+- Remove Google Drive and Dropbox storage drivers, use local driver
+- Use Lato as default UI font
+
+### Fixed
+
+- Added vite-plugin-top-level-await to silence Chrome/Chromium DevTools errors
+- Slightly lighter text color for neo-day theme
+
 ## [v3.1.0] - 2024-08-17
 
 ### Added